+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/blobs/Fire.java
@@ -43,7 +43,7 @@ public class Fire extends Blob {
 
 		boolean[] flamable = Dungeon.level.flamable;
 		int cell;
-		int fire;
+		long fire;
 		
 		Freezing freeze = (Freezing)Dungeon.level.blobs.get( Freezing.class );
