+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/blobs/Web.java
@@ -61,7 +61,7 @@ public class Web extends Blob {
 	}
 
 	@Override
-	public void seed(Level level, int cell, int amount) {
+	public void seed(Level level, int cell, long amount) {
 		super.seed(level, cell, amount);
 		level.solid[cell] = cur[cell] > 0 || (Terrain.flags[level.map[cell]] & Terrain.SOLID) != 0;
 		level.flamable[cell] = cur[cell] > 0 || (Terrain.flags[level.map[cell]] & Terrain.FLAMABLE) != 0;
