+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/hero/abilities/duelist/Feint.java
@@ -181,7 +181,7 @@ public class Feint extends ArmorAbility {
 		}
 
 		@Override
-		public int defenseSkill(Char enemy) {
+		public long defenseSkill(Char enemy) {
 			if (enemy.alignment == Alignment.ENEMY) {
 				if (enemy instanceof Mob) {
 					((Mob) enemy).clearEnemy();
