+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/GoldenMimic.java
@@ -85,7 +85,7 @@ public class GoldenMimic extends Mimic {
 	}
 
 	@Override
-	public void setLevel(int level) {
+	public void setLevel(long level) {
 		super.setLevel(Math.round(level*1.33f));
 	}
