+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/traps/Trap.java
@@ -114,7 +114,7 @@ public abstract class Trap implements Bundlable {
 	//returns the depth value the trap should use for determining its power
 	//If the trap is part of the level, it should use the true depth
 	//If it's not part of the level (e.g. effect from reclaim trap), use scaling depth
-	protected int scalingDepth(){
+	protected long scalingDepth(){
 		return (reclaimed || Dungeon.level.traps.get(pos) != this) ? Dungeon.scalingDepth() : Dungeon.depth;
 	}
